pyproject.toml
[project]
name = "geminirunner"
version = "0.1.0"
description = "CLI for common Gemini commands"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.20.0",
    "mypy>=1.16.1",
    "ruff>=0.11.13",
    "typer>=0.16.0",
]

[tool.ruff]
target-version = "py312"
line-length = 79
